relax Debug impls to only require T::Key<'a>: Debug for a particular 'a (#39)
This allows `Debug` with borrowed items to work in more circumstances.

We do have to use unsafe code, but that's okay because we have a `T: 'a`
bound. This means that the key is valid for 'a, and in any case we
discard the key immediately after calling its `Debug` impl.

It looks like we'll also have to do this for all the `IdOrdMap` mutable
getters. That will come in the future.