refactor: relax bounds on struct definitions (#122)
It's uncommon to place bounds on structs unless the trait is needed for
rustc to compute the layout.
e.g required for
[`iter::Peekable`](https://doc.rust-lang.org/beta/std/iter/struct.Peekable.html)
but not for
[`collections::HashMap`](https://doc.rust-lang.org/beta/std/collections/struct.HashMap.html).

This change means I can stop polluting my own generic structs with trait
bounds, and only put bounds where the behaviour requires it.

I think this is backwards compatible.