store hash state just once in Bi and TriHashMap (#213)
Store it just once rather than multiple times, reducing the size of Bi
and TriHashMaps. Also allows us to relax the `Clone` bound on the hasher
in a few spots, preparing for more methods to be marked as const soon.

This also allows BiHashMap and TriHashMap::with_hasher to be `const
fn`s.