add facilities to serialize as a map rather than a sequence (#215)
We serialize as a sequence by default which is correct, but in some
cases (e.g. for backwards compatibility), it is required that we
serialize as maps instead. Add serializers for this, and extend the
default deserializer to also work with maps.

Also add JSON Schema implementations that say "object" rather than
"array".