fix potential unsoundness in retain (#229)
Realized that both item and items derived from the same &mut, which
meant that they were susceptible to potential unsoundness. Fixing this
is just a matter of rearranging code so that items and item are awoken
in separate blocks.

This wasn't flagged by miri, so it's possible it was fine -- but we
should fix it anyway.